requirements.txt
# 原有依赖项保持不变
Pillow~=9.1.0
PyPDF2~=1.27.9

# 新增Flask框架依赖
Flask~=3.1.0